:brisk: refactor: 실제 dalle 요청 보내는 로직 분리
